Testes unitátios

+++ src/main/java/br/senac/sp/projeto/cineticketoficial/services/ClienteService.java
@@ -3,6 +3,9 @@ package br.senac.sp.projeto.cineticketoficial.services;
 import br.senac.sp.projeto.cineticketoficial.DTO.CadastroDTO;
 import br.senac.sp.projeto.cineticketoficial.entity.Acesso;
 import br.senac.sp.projeto.cineticketoficial.entity.Cliente;
+import br.senac.sp.projeto.cineticketoficial.exceptions.IllegalArgumentException;
+import br.senac.sp.projeto.cineticketoficial.exceptions.NullAttributesException;
+import br.senac.sp.projeto.cineticketoficial.exceptions.ResourceNotFoundException;
 import br.senac.sp.projeto.cineticketoficial.repository.ClienteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,9 +15,18 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ClienteService {
+
     private final ClienteRepository repository;
 
     public Cliente inserirCliente(CadastroDTO cadastroDTO) {
+        Cliente cliente = criarClienteAPartirDTO(cadastroDTO);
+        return this.repository.save(cliente);
+    }
+
+    private Cliente criarClienteAPartirDTO(CadastroDTO cadastroDTO) {
+        if (cadastroDTO.possuiAtributosNulos()) {
+            throw new NullAttributesException();
+        }
         Cliente cliente = new Cliente();
         cliente.setEmail(cadastroDTO.getEmail());
         cliente.setNome(cadastroDTO.getNome());
@@ -28,18 +40,25 @@ public class ClienteService {
         acesso.setSenha(cadastroDTO.getSenha());
 
         cliente.setAcesso(acesso);
-        return this.repository.save(cliente);
+        return cliente;
     }
 
+    //não é necessario ter
     public List<Cliente> buscarTodosClientes() {
         return this.repository.findAll();
     }
 
     public Cliente buscarClientePorEmail(String email) {
-        return this.repository.findById(email).orElse(null);
+        if (email == null) {
+            throw  new IllegalArgumentException("Email não pode ser Nulo");
+        }
+        return this.repository.findById(email).orElseThrow(ResourceNotFoundException::new);
     }
 
-    public Cliente excluirCliente(String email) {
+    public Cliente deletarCliente(String email) {
+        if (email == null) {
+            throw  new IllegalArgumentException("Email não pode ser Nulo");
+        }
         Cliente deleted = buscarClientePorEmail(email);
         repository.deleteById(email);
         return deleted;
